+++ pkg/controller/sync/controller.go
@@ -378,12 +378,16 @@ func (s *KubeFedSyncController) syncToClusters(fedResource FederatedResource) ut
 		runtime.HandleError(err)
 	}
 
-	statusMap := dispatcher.StatusMap()
-	return s.setPropagationStatus(fedResource, status.AggregateSuccess, statusMap)
+	collectedStatus := dispatcher.CollectedStatus()
+	return s.setPropagationStatus(fedResource, status.AggregateSuccess, &collectedStatus)
 }
 
 func (s *KubeFedSyncController) setPropagationStatus(fedResource FederatedResource,
-	reason status.AggregateReason, statusMap status.PropagationStatusMap) util.ReconciliationStatus {
+	reason status.AggregateReason, collectedStatus *status.CollectedPropagationStatus) util.ReconciliationStatus {
+
+	if collectedStatus == nil {
+		collectedStatus = &status.CollectedPropagationStatus{}
+	}
 
 	kind := fedResource.FederatedKind()
 	name := fedResource.FederatedName()
@@ -403,8 +407,11 @@ func (s *KubeFedSyncController) setPropagationStatus(fedResource FederatedResour
 	// If the underlying resource has changed, attempt to retrieve and
 	// update it repeatedly.
 	err := wait.PollImmediate(1*time.Second, 5*time.Second, func() (bool, error) {
-		if err := status.SetPropagationStatus(obj, reason, statusMap); err != nil {
+		if updateRequired, err := status.SetPropagationStatus(obj, reason, *collectedStatus); err != nil {
 			return false, errors.Wrapf(err, "failed to set the status")
+		} else if !updateRequired {
+			klog.V(4).Infof("No update necessary for %s %q propagation status", kind, name)
+			return true, nil
 		}
 
 		err := s.hostClusterClient.UpdateStatus(context.TODO(), obj)